requirements.txt
pandas
numpy
pillow
tensorflow
streamlit
pickle5